Un-exports function used only in the ETL
